% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput_CatchDataFrame.R
\name{applyAssumedCVs}
\alias{applyAssumedCVs}
\title{apply assumed CVs to catch dataframe}
\usage{
applyAssumedCVs(dfr, dfrCVs, sclF = 1)
}
\arguments{
\item{dfr}{\itemize{
\item catch dataframe with columns year,type,value
}}

\item{dfrCVs}{\itemize{
\item dataframe with columns cv, minErr, and possibly year and/or type
}}
}
\value{
dataframe with cv as column, and otherwise same columns as dfr
}
\description{
apply assumed CVs to catch dataframe
}
\details{
Adds column \code{cv} to \code{dfr} (if not already present) consistent
with values in \code{dfrCVs}. \code{dfrCVs} has columns
\itemize{
\item{cv - default value for cv}
\item{minErr - minimum standard deviation allowed}
\item{sclF - multiplicative scale factor to convert minErr to value units}
\item{year - (optional) corresponding year}
\item{type - (optional) corresponding value type (ABUNDANCE or BIOMASS)}
}
}
