% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_par_diffs.R
\name{check_par_diffs}
\alias{check_par_diffs}
\title{Check parameter values for differences between model runs}
\usage{
check_par_diffs(dn_old, dn_new)
}
\arguments{
\item{dn_old}{\itemize{
\item folder name with old model par file
}}

\item{dn_new}{\itemize{
\item folder name with new model par file
}}
}
\value{
TRUE or FALSE depending on outcome of comparisons
}
\description{
Check parameter values for differences between model runs

Function to check parameter values for differences between model runs.
}
\details{
Creates a file "rda_checks_for_different_param_values.RData" in the
parent folder of \code{dn_new} with the results of the comparison in list format.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
#--run comparison----
dirThs = getwd();
dfrRuns = tibble::tribble(~old,~new,
                         file.path("runs_20260506_Katie_v.2.30.37"),
                         file.path("runs_20260729","BBRKC_v.2.20.42b"),
                         file.path("runs_20260514_Caitlin_v.2.20.37a"),
                         file.path("runs_20260729","NSRKC_v.2.20.42b"),
                         file.path("runs_20260729","SnowCrab_v.2.20.42"),
                         file.path("runs_20260729","SnowCrab_v.2.20.42b"),
                         file.path("runs_20260707_FMEY","TannerCrab_v.2.20.41"),
                         file.path("runs_20260729","TannerCrab_v.2.20.42b")
                         );
for (rw in 4:nrow(dfrRuns)){
  #--testing: rw = 1;
  dfrRun = dfrRuns[rw,];
  dn_old = file.path(dirThs,"..",dfrRun$old,"run_all");
  dn_new = file.path(dirThs,"..",dfrRun$new,"run_all");
  if (dir.exists(dn_old)){
    if (dir.exists(dn_new)){
      cat(paste0("Comparing models ",dfrRun$old," and ",dfrRun$new,".","\n"));
      res = check_par_diffs(dn_old,dn_new);
      cat(paste0("\tComparing models ",dfrRun$old," and ",dfrRun$new,". Result: ",res,"\n\n"));
    } else {
      cat("Folder",dn_new,"not found.\n\n")
    }
  } else {
    cat("Folder",dn_old,"not found.\n\n")
  }
}
\dontshow{\}) # examplesIf}
}
