% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_SD_vars.R
\name{check_sd_vars}
\alias{check_sd_vars}
\title{Check standard errors for consistency}
\usage{
check_sd_vars(dn)
}
\arguments{
\item{dn}{\itemize{
\item full path to run folder
}}
}
\value{
TRUE or FALSE
}
\description{
Function to check the consistency of the reported standard error values in gmacs.std
and gmacs.rep1 for parameters and derived quantities.
}
\details{
Creates the file "rda_checks_for_different_sdreport_values.RData" in the parent folder
of \code{dn} with dataframes that document any inconsistent estimates/standard errors reported between
the "gmacs.std" and "gmacs.rep1" files.
}
\examples{
\dontshow{if (condition) withAutoprint(\{ # examplesIf}
# example code
#--run function----
dirThs = getwd();
# paste(list.dirs(file.path(dirThs,"..","runs_20260707_FMEY"),full.names=FALSE,recursive=FALSE),collapse=", ")
# paste(list.dirs(file.path(dirThs,"..","runs_20260729"),full.names=FALSE,recursive=FALSE),collapse=", ")
dirRuns = c(file.path("runs_20260707_FMEY",
                      c("BBRKC_v2.20.41","BBRKC_v2.20.42","NSRKC_v.2.20.37a",
                        "NSRKC_v.2.20.41","NSRKC_v.2.20.42",
                        "TannerCrab_v.2.20.37","TannerCrab_v.2.20.41","TannerCrab_v.2.20.42")),
            file.path("runs_20260729",
                      c("BBRKC_v2.20.42b", "NSRKC_v.2.20.42b", "SnowCrab_v.2.20.42a", 
                        "TannerCrab_v.2.20.42", "TannerCrab_v.2.20.42b)"))
            );
dirRuns = file.path("runs_20260729",
                      c("BBRKC_v2.20.42b", "NSRKC_v.2.20.42b", "SnowCrab_v.2.20.42b","TannerCrab_v.2.20.42b"));
dirRun = file.path("runs_20260729","BBRKC_v2.20.42b");
for (dirRun in dirRuns){
  dn = file.path(dirThs,"..",dirRun,"run_all");
  if (dir.exists(dn)&&file.exists(file.path(dn,"gmacs.std"))&&file.exists(file.path(dn,"gmacs.rep1"))){
    cat(paste0("Checking ",dirRun,"\n"));
    res = check_sd_vars(dn);
    cat("\tModel",basename(dirRun),res,"\n\n");
  } else {
    cat(paste0("Skipping ",dirRun,". Folder or gmacs.std or gmacs.rep1 does not exist.\n\n"));
  }
}
\dontshow{\}) # examplesIf}
}
