% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFitsZCsAlt.R
\name{compareFitsZCsAlt}
\alias{compareFitsZCsAlt}
\title{Plot GMACS size composition fits}
\usage{
compareFitsZCsAlt(
  dfr,
  ncols = NULL,
  subtitle = "",
  xlab = "size (mm CW)",
  ylab = "proportion",
  useBars = FALSE,
  usePins = FALSE,
  usePinsAndPts = TRUE,
  useLines = TRUE,
  usePoints = FALSE,
  plot1stObs = TRUE,
  pinSize = 0.2,
  lineSize = 0.5,
  pointSize = 0.4,
  alpha = 1
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with observed/predicted size comps from Gmacs rep1 file(s)
}}

\item{ncols}{\itemize{
\item number of panel columns for figure
}}

\item{subtitle}{\itemize{
\item (sub) title for figure (default="")
}}

\item{xlab}{\itemize{
\item x-axis label (default="size (mm CW)")
}}

\item{ylab}{\itemize{
\item y-axis label (default="proportion")
}}

\item{useBars}{\itemize{
\item use bars for observed size comps (default=FALSE)
}}

\item{usePins}{\itemize{
\item use only pins for observed size comps (default=FALSE)
}}

\item{usePinsAndPts}{\itemize{
\item use pins and points for observed size comps (default=TRUE)
}}

\item{useLines}{\itemize{
\item use lines for predicted size comps (default=TRUE)
}}

\item{usePoints}{\itemize{
\item use points for predicted size comps (default=FALSE)
}}

\item{plot1stObs}{\itemize{
\item flag to plot only first set of observations (default=TRUE)
}}

\item{pinSize}{\itemize{
\item pin size used for observed comps (default=0.2)
}}

\item{lineSize}{\itemize{
\item line size used for predicted comps (default=0.5)
}}

\item{pointSize}{\itemize{
\item size of points used for predicted comps (default=0.4)
}}

\item{alpha}{\itemize{
\item transparency (default=1)
}}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Function to plot fits to size compositions.
}
\details{
Input dataframe should represent a (possibly series of concatenated)
size composition fits dataframe(s) with columns
\itemize{
\item {case - name(s) identifying case(s) for labels}
\item {type - possible values are 'predicted' or 'observed'}
\item {y - year}
\item {z - size}
\item {val - value}
}.
}
