% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNaturalMortality.R
\name{compareNaturalMortality}
\alias{compareNaturalMortality}
\title{Create a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} object to compare tcsam02 and gmacs natural mortality rates}
\usage{
compareNaturalMortality(dfr, ylab = "M", plotZ = FALSE)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe to plot
}}

\item{ylab}{\itemize{
\item y-axis label
}}

\item{plotZ}{\itemize{
\item flag to plot size-specific mortality rates (not implemented yet)
}}
}
\value{
\link[ggplot2:ggplot2-package]{ggplot2::ggplot2} plot object
}
\description{
Function to create a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} object to compare tcsam02 and gmacs
selectivity/retention functions. Can also be used to plot either model type.
}
\details{
The input dataframe should at least have columns named "case" (model case), "y" (year),
"z" (size), and "val" (value). Bars/curves are colored by \code{case}.

The gmacs \code{case} should start with "gmacs".
}
