% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSelFcns.R
\name{compareSelFcns}
\alias{compareSelFcns}
\title{Create a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} object to compare tcsam02 and gmacs selectivity/retention functions}
\usage{
compareSelFcns(dfr, ylab = "Selectivity", sub = "")
}
\arguments{
\item{dfr}{\itemize{
\item dataframe to plot
}}

\item{ylab}{\itemize{
\item y-axis label
}}

\item{sub}{\itemize{
\item subtitle to label plot
}}
}
\value{
\link[ggplot2:ggplot2-package]{ggplot2::ggplot2} plot object
}
\description{
Function to create a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} object to compare tcsam02 and gmacs
selectivity/retention functions. Can also be used to plot either model type.
}
\details{
The input dataframe should at least have columns named "case" (model case), "y" (year),
"z" (size), and "val" (value). Curves are colored by \code{case}.
}
