% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions.R
\name{extractCohortProgression}
\alias{extractCohortProgression}
\title{Extract the cohort progression from (possibly several) GMACS and TCSAM model runs.}
\usage{
extractCohortProgression(
  lstGMACS,
  lstTCSAM02 = NULL,
  cast = "x+m+s+z",
  gmacsType = 2
)
}
\arguments{
\item{lstGMACS}{\itemize{
\item gmacs_reslst or gmacs_replst object, or NULL
}}

\item{cast}{\itemize{
\item string describing output level of aggregation (default="x+m+s+z")
}}

\item{gmacsType}{\itemize{
\item gmacs cohort progression type to extract (1: pop starts in first size bin; 2: uses recruitment -at-size distribution)
}}

\item{lstTCSAM}{\itemize{
\item tcsam02.resLst object, or NULL
}}
}
\value{
dataframe
}
\description{
Function to extract the cohort progression from (possibly several) GMACS and TCSAM model runs.
}
