% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions.R
\name{extractFitsToSurveySizeComps}
\alias{extractFitsToSurveySizeComps}
\title{Extract summary info on fits to survey size comp data from (possibly several) GMACS and TCSAM model runs.}
\usage{
extractFitsToSurveySizeComps(lstGMACS, lstTCSAM02 = NULL)
}
\arguments{
\item{lstGMACS}{\itemize{
\item gmacs_reslst or gmacs_replst object, or NULL
}}

\item{lstTCSAM}{\itemize{
\item tcsam02.resLst object, or NULL
}}
}
\value{
dataframe
}
\description{
Function to extract summary info on fits to survey size comp data from (possibly several) GMACS and TCSAM model runs.
}
