% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions.R
\name{extractMeanGrowth}
\alias{extractMeanGrowth}
\title{Extract mean growth from (possibly several) GMACS and TCSAM model runs.}
\usage{
extractMeanGrowth(lstGMACS, lstTCSAM02 = NULL)
}
\arguments{
\item{lstGMACS}{\itemize{
\item gmacs_reslst or gmacs_replst object, or NULL
}}

\item{lstTCSAM}{\itemize{
\item tcsam02.resLst object, or NULL
}}
}
\value{
dataframe
}
\description{
Function to extract mean growth from (possibly several) GMACS and TCSAM model runs.
}
