% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParamsAtBounds.R
\name{extractParamsAtBounds}
\alias{extractParamsAtBounds}
\title{Extract dataframe of parameters-at-bounds from a gmacs object.}
\usage{
extractParamsAtBounds(obj)
}
\arguments{
\item{obj}{\itemize{
\item the object from which to extract the dataframe
}}
}
\value{
a dataframe
}
\description{
Function to extract a dataframe of parameters-at-bounds from a
\code{gmacs_reslst}, \code{gmacs_repslst}, or \code{gmacs_rep1} object.
}
\details{
The returned dataframe
}
