% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractParamsInfo.R
\name{extractParamsInfo}
\alias{extractParamsInfo}
\title{Extract dataframe of parameter information from \code{gmacs_rep1} objects.}
\usage{
extractParamsInfo(obj)
}
\arguments{
\item{obj}{\itemize{
\item the object from which to extract the dataframe
}}
}
\value{
a dataframe
}
\description{
Function to extract a dataframe of parameter information from \code{gmacs_rep1}
objects packaged as a \code{gmacs_reslst}, \code{gmacs_repslst}, or a single \code{gmacs_rep1} object.
}
\details{
If results from multiple models are included in \code{obj}, the returned dataframe
has a "case" column appended to it that identifies model cases.
}
