% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions.R
\name{extractPopAbd}
\alias{extractPopAbd}
\title{Extract the population abundance time series from (possibly several) GMACS and TCSAM model runs.}
\usage{
extractPopAbd(lstGMACS, lstTCSAM02 = NULL, cast = "x+m+s+z")
}
\arguments{
\item{lstGMACS}{\itemize{
\item gmacs_reslst or gmacs_replst object, or NULL
}}

\item{cast}{\itemize{
\item string describing output level of aggregation (default="x+m+s+z")
}}

\item{lstTCSAM}{\itemize{
\item tcsam02.resLst object, or NULL
}}
}
\value{
dataframe
}
\description{
Function to extract the population abundance time series from (possibly several) GMACS and TCSAM model runs.
}
