% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions.R
\name{extractPredictedSurveyBiomass}
\alias{extractPredictedSurveyBiomass}
\title{Extract predicted survey biomass from (possibly several) GMACS and TCSAM model runs.}
\usage{
extractPredictedSurveyBiomass(
  lstGMACS,
  lstTCSAM02 = NULL,
  fleetG = NULL,
  fleetT = fleetG
)
}
\arguments{
\item{lstGMACS}{\itemize{
\item gmacs_reslst or gmacs_replst object, or NULL
}}

\item{fleetG}{\itemize{
\item gmacs name for fleet to extract
}}

\item{fleetT}{\itemize{
\item tcsam02 name for fleet to extract (default = same as gmacs name)
}}

\item{lstTCSAM}{\itemize{
\item tcsam02.resLst object, or NULL
}}
}
\value{
dataframe
}
\description{
Function to extract predicted survey biomass from (possibly several) GMACS and TCSAM model runs.
}
