% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions.R
\name{extractPrM2M}
\alias{extractPrM2M}
\title{Extract the probability of the molt-to-maturity from (possibly several) GMACS and TCSAM model runs.}
\usage{
extractPrM2M(lstGMACS, lstTCSAM02 = NULL)
}
\arguments{
\item{lstGMACS}{\itemize{
\item gmacs_reslst or gmacs_replst object, or NULL
}}

\item{lstTCSAM}{\itemize{
\item tcsam02.resLst object, or NULL
}}

\item{cast}{\itemize{
\item string describing output level of aggregation (default="x+m+s+z")
}}
}
\value{
dataframe
}
\description{
Function to extract the molt-to-maturity from (possibly several) GMACS and TCSAM model runs.
}
