% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRep1Results.R
\name{extractRep1Results}
\alias{extractRep1Results}
\title{Extract an object from \code{gmacs_rep1} objects.}
\usage{
extractRep1Results(obj, objname)
}
\arguments{
\item{obj}{\itemize{
\item a gmacs_reslst, gmacs_repslst, or gmacs_rep1 object
}}

\item{objname}{\itemize{
\item name of object to extract
}}
}
\value{
If the object to extract is a dataframe, then the returned is a dataframe
with an appended column 'case' indicating the
originating model case. Otherwise a list of extracted objects named by case.
}
\description{
Function to extract an object from \code{gmacs_rep1} objects.
}
\details{
If the object to extract is a dataframe, then the returned is a dataframe
with an appended column 'case' indicating the
originating model case. Otherwise a list of extracted objects named by case.
}
