% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSelFcns.R
\name{extractSelFcns}
\alias{extractSelFcns}
\title{Extract selectivity/retention functions from a gmacs and a tcsam02 model run}
\usage{
extractSelFcns(resGMACS, resTCSAM02 = NULL)
}
\arguments{
\item{resGMACS}{\itemize{
\item 
}}

\item{resTCSAM02}{\itemize{
\item 
}}
}
\value{
dataframe
}
\description{
Funciton to extract selectivity/retention functions from a gmacs and a tcsam02 model run.
}
\details{
Return a dataframe in TCSAM02 standard canonical format.
}
