% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestJitterRun.R
\name{findBestJitterRun}
\alias{findBestJitterRun}
\title{Determine best GMACS jitter run among a series of runs}
\usage{
findBestJitterRun(fldrs, verbose = FALSE)
}
\arguments{
\item{fldrs}{\itemize{
\item character vector of folder names to search over
}}

\item{verbose}{\itemize{
\item flag to print plot of objective function results
}}
}
\value{
list with elements \code{dfrJtr}, \code{plt},\code{minObjFun}, and \code{maxGrad}.
}
\description{
Function to determine best GMACS jitter run among a series of runs.
}
\details{
The "jitterResults.csv" file in each folder is read in, concatenated
into a single dataframe, and arranged by order of the objective function value
for each run. Details, including the folder name, of the "best" model are
given in a message. A count of the number of model runs with "similar"
objective function values (closer than 0.001 to the minimum) is also included.
In addition, a plot is provided of maximum gradient vs. objective function
value for all model runs.
}
