% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getDefaultRunCmds}
\alias{getDefaultRunCmds}
\title{Get default run commands}
\usage{
getDefaultRunCmds()
}
\value{
list with default input values for \link{getRunCommands}
}
\description{
Function to get default run commands.
}
\details{
Use the returned list (after modification, as necessary) as
input to \link{getRunCommands}:\cr

\if{html}{\out{<div class="sourceCode">}}\preformatted{deflst = getDefaultRunCmds(); 
#..update defaults...
runcmds = do.call(deflst,getRunCommands);
}\if{html}{\out{</div>}}
}
