% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions2.R
\name{getFitsToIndexTimeSeries}
\alias{getFitsToIndexTimeSeries}
\title{Get fits to index time series}
\usage{
getFitsToIndexTimeSeries(lstGs, lstT = NULL, type = "biomass")
}
\arguments{
\item{lstGs}{\itemize{
\item list with named GMACS repslst objects from model runs
}}

\item{lstT}{\itemize{
\item TCSAM02 reslst object
}}

\item{type}{\itemize{
\item index type to extract ('biomass' or 'abundance')
}}
}
\value{
dataframe
}
\description{
Functio to get fits to index time series from GMACS and TCSAM02 models.
}
\details{
For GMACS results, the ln-scale \code{sdobs} is calculated from the \code{cv},
the \code{zscore} is taken from the pearson's residual (\code{prsn_res}), NLLs for
individual observations/predictions are recalculated assuming a lognormal
likelihood and the associated \code{sdobs} and \code{zscore}. the \code{residual} is calculated
on the ln scale as $log(observed) - log(predicted)$.

For TCSAM02 results, the \code{residual} is calculated as above, the \code{cv} is calculated from the
\code{sdobs}, the NLL is recalculated for individual observations consistent with that
for GMACS above.

For all, the 90\% CIs are calculated for the observations based on the CVs.
}
