% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions2.R
\name{getFitsToZCs}
\alias{getFitsToZCs}
\title{Get fits to size comps from GMACS and (possibly) TCSAM02 models}
\usage{
getFitsToZCs(lstGs, lstT = NULL)
}
\arguments{
\item{lstGs}{\itemize{
\item list with named GMACS repslst objects from model runs
}}

\item{lstT}{\itemize{
\item TCSAM02 reslst object
}}

\item{type}{\itemize{
\item index type to extract ('biomass' or 'abundance')
}}
}
\value{
dataframe
}
\description{
Function to get fits to size comps from GMACS and (possibly) TCSAM02 models.
}
