% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListOfRep1Objects.R
\name{getListOfRep1Objects}
\alias{getListOfRep1Objects}
\title{Create a list of rep1 objects}
\usage{
getListOfRep1Objects(pns, verbose = FALSE)
}
\arguments{
\item{pns}{\itemize{
\item (possibly named) vector of paths to gmacs.rep1 files.
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
named list of \code{rep1} objects.
}
\description{
Function to create a list of rep1 objects.
}
\details{
If \code{pns} is a named vector, then the names are used as
names in the return list. Otherwise, names for the returned list
are derived from the parent folder of each gmacs.rep1 file.
}
