% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelConfigurations.R
\name{getModelConfigurations}
\alias{getModelConfigurations}
\title{Extract model configuration info}
\usage{
getModelConfigurations(obj)
}
\arguments{
\item{obj}{\itemize{
\item the object to extract the info from
}}
}
\value{
list of lists
}
\description{
Function to extract model configuration info from a
gmacs_reslst, gmacs_repslst, or gmacs_rep1 object.
}
\details{
list of lists, one for each model
}
