% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getRunCommands}
\alias{getRunCommands}
\title{Generate run commands for a GMACS model run}
\usage{
getRunCommands(
  inp_fns = NULL,
  exe = "gmacs",
  path2exe = ".",
  path2dat = ".",
  path2pin = ".",
  pinFile = NULL,
  os = NULL,
  hess = FALSE,
  minPhase = NULL,
  maxPhase = NULL,
  mcmc = FALSE,
  mc.N = 1e+06,
  mc.save = 1000,
  mc.scale = 1000,
  jitter = FALSE,
  iseed = NULL,
  cout = "chk.rep",
  fullClean = FALSE,
  cleanup = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{inp_fns}{\itemize{
\item name(s) of data, ctl, prj files (default = NULL: must supply one or all)
}}

\item{exe}{\itemize{
\item name of gmacs executable (without extension) (default='gmacs')
}}

\item{path2exe}{\itemize{
\item path to gmacs executable (default = ".")
}}

\item{path2dat}{\itemize{
\item path to folder with input dat, data, ctl, and prj files (default = ".")
}}

\item{path2pin}{\itemize{
\item path to folder with pin file (default = ".")
}}

\item{pinFile}{\itemize{
\item the name of the pin file used to initialize the parameters
}}

\item{os}{\itemize{
\item 'win', 'mac', 'osx', or 'linux'
}}

\item{hess}{\itemize{
\item flag (T/F) to calculate the hessian
}}

\item{minPhase}{\itemize{
\item start phase (or NULL) for minimization calculations
}}

\item{maxPhase}{\itemize{
\item last phase (or NULL) for minimization calculations
}}

\item{mcmc}{\itemize{
\item flag (T/F) to do mcmc calculations
}}

\item{mc.N}{\itemize{
\item number of mcmc iterations to do
}}

\item{mc.save}{\itemize{
\item number of iterations to skip when saving mcmc calculations
}}

\item{mc.scale}{\itemize{
\item number of iterations to adjust scale for mcmc calculations
}}

\item{jitter}{\itemize{
\item flag (T/F) to use jitter initial values
}}

\item{iseed}{\itemize{
\item value for random number seed (or 0 to use start time)
}}

\item{cout}{\itemize{
\item filename to direct commandline output to (or NULL to direct it to console)
}}

\item{fullClean}{\itemize{
\item flag to clean up almost all files (use when making multiple jitter runs)
}}

\item{cleanup}{\itemize{
\item flag (T/F) to clean up unnecessary files
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info for this function
}}
}
\value{
string for shell script or batch file to run gmacs model.
}
\description{
Function to generate a script to make a GMACS2 model run.
}
\details{
. All files in \code{path2dat} are copied to the "current" folder. If the
pinFile to be used is in this folder, it will be copied into the "current" folder and
only the file name (e.g., 'gmacs.pin') needs to be given. If \code{pinFile} is NULL,
the model is initialized from the ctl file.
If \code{cleanup} is TRUE, then .bar, .b0*, .p0*, .r0*, variance,
and fimn.log files are deleted.\cr
}
