% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSelFcns.R
\name{getUniqSelFcns}
\alias{getUniqSelFcns}
\title{Get unique sel/ret functions from a set of model cases and years}
\usage{
getUniqSelFcns(dfr)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe of sel/ret functions from \code{\link[=extractSelFcns]{extractSelFcns()}}
}}
}
\description{
Function to get unique sel functions from a set of model cases and years.
}
\details{
Extracts unique sel/ret functions and labels them by model "case" and years.
}
