% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions2.R
\name{plotDataToZCs}
\alias{plotDataToZCs}
\title{Compare size composition data}
\usage{
plotDataToZCs(dfr, points = "g1")
}
\arguments{
\item{dfr}{\itemize{
\item dataframe created using \link{getFitsToZCs}.
}}
}
\value{
\href{ggplot}{ggplot2} object
}
\description{
Function to compare size composition.
}
