% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions2.R
\name{plotFitsToIndexTSs}
\alias{plotFitsToIndexTSs}
\title{Compare fits to index time series}
\usage{
plotFitsToIndexTSs(dfr, f_ = "NMFS")
}
\arguments{
\item{dfr}{\itemize{
\item dataframe created using \link{getFitsToIndexTimeSeries}.
}}

\item{f_}{\itemize{
\item fleet name (as specific for GMACS runs)
}}
}
\value{
\href{ggplot}{ggplot2} object
}
\description{
Function to compare fits to index time series.
}
