% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions2.R
\name{plotFitsToZCs}
\alias{plotFitsToZCs}
\title{Compare fits to size composition data}
\usage{
plotFitsToZCs(dfr)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe created using \link{getFitsToZCs}.
}}
}
\value{
\href{ggplot}{ggplot2} object
}
\description{
Function to compare fits to size composition.
}
