% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions2.R
\name{plotResidualsToZCs}
\alias{plotResidualsToZCs}
\title{Compare residuals for fits to size composition data}
\usage{
plotResidualsToZCs(dfr)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe created using \link{getFitsToZCs}.
}}
}
\value{
\href{ggplot}{ggplot2} object
}
\description{
Function to compare residuals for fits to size composition.
}
