% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGmacsAllout.R
\name{readGmacsAllout}
\alias{readGmacsAllout}
\title{Read a Gmacsall.out file}
\usage{
readGmacsAllout(fn, verbose = FALSE)
}
\arguments{
\item{fn}{\itemize{
\item path to the Gmacsall.out file
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a named list of class \code{gmacs_allout}
}
\description{
Read a Gmacsall.out file and return a list of report objects.
}
\details{
The returned object has class \code{gmacs_allout}` and inherits from \code{list}.

The function uses a counter, \code{iln}, in the enclosing environment. If it does not exist already,
it will be created and set to 1 initially. If it already exists when this function is called,
it will  be reset to 1 initially.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
  fn = "../../ModelRuns/TannerCrab24_01/Gmacsall.out";
  rep = readGmacsAllout(fn,verbose=FALSE);
\dontshow{\}) # examplesIf}
}
