% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGmacsAllStdFile.R
\name{readGmacsAllStdFile}
\alias{readGmacsAllStdFile}
\title{Create a \code{gmacs_allstd} dataframe from GMACS Gmacsall.std files}
\usage{
readGmacsAllStdFile(stdfn = NULL, verbose = FALSE)
}
\arguments{
\item{stdfn}{\itemize{
\item path/filename of Gmacsall.std file (or list or vector of such)
}}

\item{verbose}{\itemize{
\item flag to print diagnostics
}}
}
\value{
a \code{gmacs_allstd} object (a \link[tibble:tibble]{tibble::tibble} dataframe of class 'gmacs_allstd') corresponding to the
std file(s)
}
\description{
Function to create a \code{gmacs_allstd} dataframe from Gmacsall.std files.
}
\details{
The returned tibble as columns
\itemize{
\item{case - model case name (taken from the list or vector element name) or number. Value is 'gmacs' if only a single unnamed filename is given.}
\item{index - index of the parameter value in the std file}
\item{param - name of the parameter as given in the std file}
\item{name - parsed parameter name, dropping any \verb{[...]} index notation}
\item{idx - the \code{...} within any \verb{[...]} in the parameter name, or NA if none}
\item{est - parameter estimate or convergence quantity value}
\item{std - estimated (delta-method) standard error}
}
}
