% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGmacsRep1.R
\name{readGmacsRep1}
\alias{readGmacsRep1}
\title{Read a Gmacs gmacs.rep1 file}
\usage{
readGmacsRep1(fn, verbose = FALSE)
}
\arguments{
\item{fn}{\itemize{
\item path to gmacs.rep1 file
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a named list of class \code{gmacs_rep1}
}
\description{
Read a Gmacs .rep1 file and return a list of report objects.
}
\details{
The returned object has class \code{gmacs_rep1}` and inherits from \code{list}.

The function uses a counter, \code{iln}, in the enclosing environment. If it does not exist already,
it will be created and set to 1 initially. If it already exists when this function is called,
it will  be reset to 1 initially.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# example code
  fn = "../../ModelRuns/TannerCrab24_01/gmacs.rep1";
  rep = readGmacsRep1(fn,verbose=FALSE);
\dontshow{\}) # examplesIf}
}
