% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readModelResults.R
\name{readModelResults}
\alias{readModelResults}
\title{Read gmacs results files from a (possibly several) results folder(s)}
\usage{
readModelResults(fldrs, verbose = FALSE)
}
\arguments{
\item{fldrs}{\itemize{
\item a vector of folder names, each with the results from a single model run
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
a \code{gmacs_reslst} object: a named list (see [@details])
}
\description{
Function to read a set of gmacs model results from a (possibly several) model run folder(s).
}
\details{
A \code{gmacs_reslst} object is a list with elements
\itemize{
\item{dfrPars - a \code{gmacs_par} dataframe object with parameter and convergence info from the par file(s), idenitified by \code{case}}
\item{dfrStds - a \code{gmacs_std} dataframe object, with parameter and sd-variable estimates and uncertainty info from the std file(s), idenitified by \code{case}}
\item{dfrAllStds - a \code{gmacs_allstd} dataframe object with parameter and sd-variable estimates and uncertainty info from the Gmacsall.std file(s), identified by case}
\item{repsLst - a \code{gmacs_repslst} object, identified by \code{case}: a list of gmacs_rep1 objects (see \code{\link[=readGmacsRep1]{readGmacsRep1()}})}
\item{outsLst - a \code{gmacs_outslst} object, identified by \code{case}: a list of gmacs_allout objects (see \code{\link[=readGmacsAllout]{readGmacsAllout()}})}
}
If \code{fldrs} is a named vector, then the returned
elements of \code{repLsts} are correspondingly named. If not, the folder names are used. For
the returned dataframes, the \code{case} column contains the relevant name.
}
