% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runJitter.R
\name{runJitter}
\alias{runJitter}
\title{Function to run GMACS multiple times using jittered initial parameter values.}
\usage{
runJitter(
  inp_fns = NULL,
  pinFile = NULL,
  exe = "gmacs",
  path2exe = ".",
  path2dat = ".",
  path2pin = path2dat,
  path2out = ".",
  numRuns = 3,
  minPhase = 1,
  onlyEvalJitter = FALSE,
  out.csv = "jitterResults.csv",
  calcOFL = FALSE,
  calcOFLJitter = FALSE,
  saveResults = FALSE,
  cout = "chk.rep",
  cleanup = TRUE,
  keepFiles = c("tmp.sh", "gmacs.par", "jitter.csv"),
  cleanupAll = FALSE,
  test = TRUE
)
}
\arguments{
\item{inp_fns}{\itemize{
\item name(s) of data, ctl, prj files (default = NULL: must supply one or all)
}}

\item{pinFile}{\itemize{
\item pin file to set values for variables with phase < 0 (not estimated) (default: NULL)
}}

\item{exe}{\itemize{
\item name of gmacs executable (without extension) (default='gmacs')
}}

\item{path2exe}{\itemize{
\item path to model executable (default: '.')
}}

\item{path2dat}{\itemize{
\item path to model dat, ctl, and prj files (default: '.')
}}

\item{path2pin}{\itemize{
\item path to pin file (default: \code{path2dat})
}}

\item{path2out}{\itemize{
\item path to top folder for model output (default: '.')
}}

\item{numRuns}{\itemize{
\item number of jitter runs to make
}}

\item{minPhase}{\itemize{
\item phase in which to start optimization
}}

\item{onlyEvalJitter}{\itemize{
\item flag (T/F) to only evaluate a (previous) set of jitter runs, not make new runs
}}

\item{out.csv}{\itemize{
\item filename for jittered results
}}

\item{calcOFL}{\itemize{
\item does nothing at present (default=FALSE)
}}

\item{calcOFLJitter}{\itemize{
\item does nothing at present (default=FALSE)
}}

\item{saveResults}{\itemize{
\item (default=FALSE)
}}

\item{cout}{\itemize{
\item filename to direct commandline output to (or NULL to direct it to console)
}}

\item{cleanup}{\itemize{
\item T/F to clean up SOME model output files after each run
}}

\item{keepFiles}{\itemize{
\item vector of file names to keep, not clean up, after model run
}}

\item{cleanupAll}{\itemize{
\item T/F to clean up ALMOST ALL model output files after each run
}}

\item{test}{\itemize{
\item T/F to run as a test
}}
}
\value{
\itemize{
\item list w/ 4 elements:
\itemize{
\item{imx  - index of (1st) smallest value for the objective function}
\item{seed - seed resulting in the smallest objective function}
\item{par  - dataframe with par results from run w/ smallest objective function}
\item{objFuns - vector of objective function values from all model runs}
\item{parList - list of par dataframes for each model run}
}
}
}
\description{
This functions runs a GMACS model multiple times, jittering the
initial starting values to assess model convergence.
}
\details{
For each model run, this function creates a shell script ('./tmp.sh') in the
current working directory and uses it to run the ADMB version of the GMACS model.
Initial model parameters with phase > 0 are jittered based on the system clock time as a seed
to the random number generator. If a pin file is given (\code{pinFile}), the values for
model parameters that are \emph{not} estimated are taken from that file, otherwise they are
taken from the initial values specified in the ctl file. The seed and final objective function value are
saved for each model run in a csv file (the value of out.csv).

When all the models requested have been run,
the function determines the seed associated with the 1st model run that yielded
the smallest value for the objective function.
}
