% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput_CatchDataFrame.R
\name{writeInput_CatchDataFrame}
\alias{writeInput_CatchDataFrame}
\title{Write fishery catch dataframe to a connection in GMACS alternative 1 input format}
\usage{
writeInput_CatchDataFrame(
  conn = stdout(),
  dfr = NULL,
  fleet_name,
  fleet_names = NULL,
  gear = c("all", "fixed", "trawl", "crab pot"),
  sex = c("male", "female", "undetermined"),
  maturity = c("immature", "mature", "undetermined"),
  shell = c("new shell", "old shell", "undetermined"),
  catchType = c("RETAINED", "DISCARD", "TOTAL"),
  discard_mortality = 0,
  unitsType = c("ABUNDANCE", "BIOMASS"),
  unitsIn = ifelse(toupper(unitsType[1]) == "ABUNDANCE", c("ONES", "MILLIONS"), c("KG",
    "THOUSANDS_MT", "MILLIONS_LBS", "MT", "LBS")),
  unitsOut = ifelse(toupper(unitsType[1]) == "ABUNDANCE", c("MILLIONS", "ONES"),
    c("THOUSANDS_MT", "MILLIONS_LBS", "MT", "KG", "LBS"))
)
}
\arguments{
\item{dfr}{: aggregate catch dataframe to write to connection}

\item{fleet_name}{: name of fleet}

\item{fleet_names}{: vector of fleet_names (order corresponds to input order to GMACS)}

\item{sex}{: sex}

\item{catchType}{: catch type ("RETAINED","DISCARD",or "TOTAL")}

\item{discard_mortality}{: assumed discard (handling) mortality rate}

\item{unitsType}{: units type ("ABUNDANCE" or "BIOMASS")}

\item{unitsIn}{: input catch units (on of "ONES","MILLIONS","KG","THOUSANDS_MT","MILLIONS_LBS","MT","LBS")}

\item{unitsOut}{: output catch units (should be "MILLIONS" or "THOUSANDS_MT")}

\item{con}{: connection to use (default=stdout)}
}
\value{
Invisibly returns the connection \code{con} to allow piping.
}
\description{
Function to write a catch dataframe to a connection in GMACS alternative 1 input format.
}
\details{
The input dataframe should have the following column names
\itemize{
\item{year - observation year}
\item{sex - string indicating sex}
\item{season - integer indicating season of observation}
\item{value - observation value}
\item{cv - observation cv}
\item{type - units type ("ABUNDANCE" or "BIOMASS")}
\item{effort - fishing effort}
}
}
