% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput_ZCsDataFrame.R
\name{writeInput_ZCsDataFrame}
\alias{writeInput_ZCsDataFrame}
\title{Write size comps dataframe to a connection in GMACS input format}
\usage{
writeInput_ZCsDataFrame(
  conn = stdout(),
  dfr = NULL,
  fleet_name,
  fleet_names = NULL,
  gear = c("all", "fixed", "trawl", "pot"),
  sex = c("MALE", "FEMALE", "ALL"),
  catchType = c("RETAINED", "DISCARD", "TOTAL"),
  unitsIn = c("ONES", "MILLIONS"),
  unitsOut = "MILLIONS"
)
}
\arguments{
\item{fleet_name}{: name of fleet}

\item{fleet_names}{: vector of fleet_names (order corresponds to input order to GMACS)}

\item{sex}{: sex}

\item{catchType}{: catch type ("RETAINED","DISCARD","TOTAL")}

\item{unitsIn}{: input catch units (on of "ONES","MILLIONS")}

\item{unitsOut}{: output catch units (should be "MILLIONS" for input file)}

\item{con}{: connection to use (default=stdout)}

\item{dfrZCs}{: size comps dataframe to write to connection}
}
\value{
Invisibly returns the connection \code{con} to allow piping.
}
\description{
Function to write a size comps dataframe to a connection in GMACS input format.
}
\details{
The input dataframe should have the following column names
\itemize{
\item{year - observation year}
\item{sex - string indicating sex}
\item{season - integer indicating season of observation}
\item{value - observation value}
}
}
