% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_Functions.R
\name{extractFisheryFs}
\alias{extractFisheryFs}
\title{Extract fishing mortality rates from (possibly several) GMACS and TCSAM model runs.}
\usage{
extractFisheryFs(lstGMACS, lstTCSAM02 = NULL, seasons = 2)
}
\arguments{
\item{lstGMACS}{\itemize{
\item gmacs_reslst or gmacs_replst object, or NULL
}}

\item{seasons}{\itemize{
\item integer vector seasons to extract fishery info for (or "all"; default=2)
}}

\item{lstTCSAM}{\itemize{
\item tcsam02.resLst object, or NULL
}}
}
\value{
dataframe with TCSAM02 canonical columns + \code{season}.
}
\description{
Function to extract fishing mortality rates from (possibly several) GMACS and TCSAM model runs.
}
\details{
The dataframe column \code{season} will be NA for TCSAM02 model results.
}
