% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNaturalMortality.R
\name{extractNaturalMortality}
\alias{extractNaturalMortality}
\title{Extract natural mortality rates from a gmacs and a tcsam02 model run}
\usage{
extractNaturalMortality(resGMACS, resTCSAM02 = NULL)
}
\arguments{
\item{resGMACS}{\itemize{
\item a gmacs_reslst or gmacs_rep1 object
}}

\item{resTCSAM02}{\itemize{
\item 
}}
}
\value{
dataframe
}
\description{
Function to extract natural mortality rates from a gmacs and a tcsam02 model run.
}
\details{
Returns a dataframe in TCSAM02 standard canonical format.
}
