% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRep1Results.R
\name{extractRep1Results}
\alias{extractRep1Results}
\title{Extract an object from a \code{rep1} list object.}
\usage{
extractRep1Results(lst, objname)
}
\arguments{
\item{lst}{\itemize{
\item named list of rep1 objects
}}

\item{objname}{\itemize{
\item name of object to extract
}}
}
\value{
If the object to extract is a dataframe, the returned is a dataframe
with an appended column 'case' indicating the
originating model case. Otherwise a list of extracted objects named by case.
}
\description{
Function to extract an object from a \code{rep1} list object.
}
