% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGmacsRep1.R
\name{readGmacsRep1}
\alias{readGmacsRep1}
\title{Read a Gmacs gmacs.rep1 file}
\usage{
readGmacsRep1(fn, verbose = FALSE)
}
\arguments{
\item{fn}{\itemize{
\item path to gmacs.rep1 file
}}

\item{verbose}{\itemize{
\item flag to print diagnostic info
}}
}
\value{
a named list of class \code{rep1}
}
\description{
Read a Gmacs .rep1 file and return a list of report objects.
}
\details{
Must define \code{iln=1} in global environment before calling ths function.
The returned object has class \code{rep1}` and inherits from \code{list}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
  fn = "../../ModelRuns/TannerCrab24_01/gmacs.rep1";
  iln = 1; #--must set iln=1 to start!!
  rep = readGmacsRep1(fn,verbose=FALSE);
\dontshow{\}) # examplesIf}
}
