% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParFile.R
\name{readParFile}
\alias{readParFile}
\title{Create a \code{gmacspar} dataframe from an ADMB par file.}
\usage{
readParFile(parfn = NULL)
}
\arguments{
\item{parfn}{= filename of par file}
}
\value{
a \code{gmacspar} object (a \link{tibble} dataframe of class 'gmacspar') corresponding to the par file (or NULL if file does not exist)
}
\description{
Function to create a \code{gmacspar} dataframe from an ADMB par file.
}
\details{
The returned tibble as columns
\itemize{
\item{name - parsed parameter name (or 'number of parameters','objective function','max gradient')}
\item{value - parameter estimate or convergence quantity value
}
}
The ADMB convergence information is given in the first three rows of the tibble,
associated with the names 'number of parameters','objective function',and 'max gradient'.

Note that this is not the same format as returned by \code{\link[gmr:readGMACSpar]{gmr::readGMACSpar()}}, which
contains more structured information based on the model run's dat and ctl files.
}
