% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGMACS.R
\name{runGMACS}
\alias{runGMACS}
\title{Run a GMACS model}
\usage{
runGMACS(
  runpath = ".",
  runCmds = list(path2exe = ".", path2dat = ".", pinFile = NULL, os = NULL, hess = FALSE,
    minPhase = NULL, maxPhase = NULL, mcmc = FALSE, mc.N = 1e+06, mc.save = 1000,
    mc.scale = 1000, jitter = FALSE, iseed = NULL, fullClean = FALSE, cleanup = TRUE,
    verbose = FALSE),
  test = FALSE
)
}
\arguments{
\item{runpath}{\itemize{
\item path to folder in which to run model
}}

\item{runCmds}{\itemize{
\item list of inputs to \code{\link[=getRunCommands]{getRunCommands()}}
}}
}
\value{
NULL (if \code{test}==TRUE), or a \code{gmacspar}-class dataframe.
}
\description{
Function to run a GMACS model.
}
\details{
If \code{test == FALSE} and the model successfully converges, the returned
value is a dataframe of class \code{gmacspar} from \code{\link[=readParFile]{readParFile()}}. In addition, if
jittering was done, the random number generator seed is prepended as the first
row.
}
