% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput_RelativeAbundanceDataFrame.R
\name{writeInput_RelativeAbundanceDataFrame}
\alias{writeInput_RelativeAbundanceDataFrame}
\title{Write a relative abundance dataframe to a connection in GMACS alternative 1 input format}
\usage{
writeInput_RelativeAbundanceDataFrame(
  conn = stdout(),
  dfr = NULL,
  fleet_name,
  sex = c("male", "female", "undetermined"),
  maturity = c("immature", "mature", "undetermined"),
  shell = c("new shell", "old shell", "undetermined"),
  indexType = c("selectivity", "selectivity+retention"),
  unitsType = c("abundance", "biomass"),
  unitsIn = ifelse(toupper(unitsType[1]) == "ABUNDANCE", c("ONES", "MILLIONS"), c("KG",
    "THOUSANDS_MT", "MILLIONS_LBS", "MT", "LBS")),
  unitsOut = ifelse(toupper(unitsType[1]) == "ABUNDANCE", c("MILLIONS", "ONES"),
    c("THOUSANDS_MT", "MILLIONS_LBS", "MT", "KG", "LBS"))
)
}
\arguments{
\item{dfr}{: an abundance indices dataframe to write to connection}

\item{fleet_name}{: name of fleet}

\item{sex}{: "male","female",or "undetermined"}

\item{maturity}{\itemize{
\item "immature","mature", or "undetermined"
}}

\item{shell}{\itemize{
\item "new shell","old shell",or "undetermined"
}}

\item{indexType}{: index type ("sel" \link{selectivity} or "sel+ret" \link{selectivity+retention})}

\item{unitsType}{: units type ("abundance" or "biomass")}

\item{unitsIn}{: input catch units (on of "ONES","MILLIONS","KG","THOUSANDS_MT","MILLIONS_LBS","MT","LBS")}

\item{unitsOut}{: output catch units (should be "MILLIONS" or "THOUSANDS_MT")}

\item{con}{: connection to use (default=stdout)}
}
\value{
Invisibly returns the connection \code{con} to allow piping.
}
\description{
Function to write a relative abundance dataframe to a connection in GMACS alternative 1 input format.
}
\details{
The input dataframe should have the following column names
\itemize{
\item{q_index - index for q parameter}
\item{year - observation year}
\item{season - integer indicating season of observation}
\item{value - observation value}
\item{cv - observation cv}
\item{timing - relative timing within season}
}
}
